gitignore: Added Visual Studio, CodeBlocks and QtCreator project files
We are using CMake to build the project. Since everyone will be using
the development environment she or he prefers, there is the risk one is
starting a task on Visual Studio, the other one a task on Debian with
CodeBlocks, another one starts a task with Emacs and of course does not
make changes at the project files. And then, some smart mind using
CodeBlocks on Windows says "oh cool, there's a CodeBlocks project file"
and may start with an obsolete project state.

Everyone building and developing UFM is responsible for setting up the
build or development environment the way he or she is used to.

If there are contributors using other IDEs creating project or temporary
files which are not mentioned in the .gitignore file yet, please feel
free to add them and write a tutorial how you set up the project with
this IDE! It could be helpful for other users once ;)